+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [#354 - Product search in elasticsearch](https://github.com/shopsys/shopsys/pull/354)
     - elasticsearch docker container: volumes are now set in order to preserve data when the container is shut down
     - added Symfony commands for creating and removing Elasticsearch indexes
+    - added Symfony command for exporting all products to Elasticsearch
 
 #### Fixed
 - [#260 - JS validation: dynamically added form inputs are now validated](https://github.com/shopsys/shopsys/pull/260)